Added setup for recurring task that checks for filter jobs to process
I use node-cron to setup the recurring task in Express and once the task
identifies a filter job to process, it uses node-fetch to talk to the
NH Algolia API (https://hn.algolia.com/api). I did some basic smoke
testing for a few filter jobs and was able to generate filter results
but still need to refine the job selection logic and add error handling.